refactor(type): change order type from int -> float
- server không quan tâm tới hiển thị, client tự xử lý order

+++ src/Grimoire.Application/Dto/Book/ChapterResponseDto.cs
@@ -6,7 +6,7 @@ using Segment;
 
 public class ChapterResponseDto : ITimestampedDto {
 	public string VolumeId { get; init; } = string.Empty;
-	public int Order { get; init; }
+	public float Order { get; init; }
 	public string Title { get; init; } = string.Empty;
 	public List<SegmentDto> Content { get; init; } = [];
 	public List<FootnoteSegmentDto> Footnotes { get; init; } = [];
